pyproject.toml
[project]
name = "vote-with-your-feet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pytest>=9.0.1",
    "pytest-describe>=2.0.1",
    "flask>=3.0.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-socketio>=5.3.6",
    "flask-cors>=4.0.0",
    "python-dotenv>=1.0.0",
]
